refactor: freeway best-practice alignment + frontend hardening
Backend: Orm IoC injection, DTO validation, named SQL params, FQN cleanup,
count() fixes, Admin RBAC/permissions, SimpleCache, test coverage (3/3)

Frontend: XSS fix, image grid plugin, types, error handling, dark mode,
auth redirect, prop mutation fix, memory leak fix, aria labels

+++ point-domain/src/main/java/com/jujin/point/domain/entity/ArticleTag.java
@@ -10,10 +10,10 @@ public record ArticleTag(
     @Id @Generated Long id,
     @Column(nullable = false) Long articleId,
     @Column(nullable = false) Long tagId,
-    @Column(nullable = false) long status,
+    @Column(nullable = false) int status,
     @Column(nullable = false) long createTime
 ) {
-    public ArticleTag(Long articleId, Long tagId, long status, long createTime) {
+    public ArticleTag(Long articleId, Long tagId, int status, long createTime) {
         this(null, articleId, tagId, status, createTime);
     }
 }
